+++ Sources/VCore/Extensions/SwiftUI/Image/Image.InitWithData.swift
@@ -26,7 +26,7 @@ extension Image {
         guard
             let uiImage: UIImage = .init(data: data)
         else {
-            VCoreFatalError("Could not initialize `UIImage` from the specified data.")
+            VCoreFatalError("Could not initialize 'UIImage' from the specified data.")
         }
         
         self.init(uiImage: uiImage)
@@ -35,7 +35,7 @@ extension Image {
         guard
             let nsImage: NSImage = .init(data: data)
         else {
-            VCoreFatalError("Could not initialize `NSImage` from the specified data.")
+            VCoreFatalError("Could not initialize 'NSImage' from the specified data.")
         }
         
         self.init(nsImage: nsImage)