+++ Sources/VCore/Extensions/SwiftUI/View/View.FitToAspect.swift
@@ -11,11 +11,11 @@ import SwiftUI
 extension View {
     /// Constrains this `View`'s dimensions to the specified aspect ratio without stretching it.
     ///
-    ///     Image(named: "SomeImage")
+    ///     Image("SomeImage")
     ///         .resizable()
     ///         .fitToAspect(1, contentMode: .fit)
     ///         .background(content: { Color.black })
-    ///         .frame(dimension: 100)
+    ///         .frame(dimension: 200)
     ///
     public func fitToAspect(
         _ aspectRatio: Double,