+++ Sources/VCore/Helpers/Architectural Pattern Helpers/UIKit/UICollectionViewProtocols.swift
@@ -132,7 +132,7 @@ extension UICollectionView {
         guard
             let cell = dequeueReusableCell(withReuseIdentifier: parameter.reuseID, for: indexPath) as? any ConfigurableUICollectionViewCell
         else {
-            VCoreFatalError("Unable to dequeue a cell with identifier `\(parameter.reuseID)`")
+            VCoreFatalError("Unable to dequeue a cell with identifier '\(parameter.reuseID)'")
         }
         
         cell.configure(parameter: parameter)